+++ game/source/ai/activities.hpp
@@ -3,9 +3,9 @@
 struct s_posture_info
 {
 	e_ai_posture posture;
-	long posture_name;
+	int32 posture_name;
 	e_posture_category category;
-	short involuntary_vocalization_index;
+	int16 involuntary_vocalization_index;
 	bool use_posture_physics;
 	bool disallow_targeting;
 	bool debilitated;
@@ -13,9 +13,9 @@ struct s_posture_info
 static_assert(sizeof(s_posture_info) == 0x14);
 
 extern e_ai_posture const(&g_activity_posture_variants)[k_activity_count][k_max_activity_variants];
-extern long const(&g_activity_names)[k_activity_count];
+extern int32 const(&g_activity_names)[k_activity_count];
 extern s_posture_info const(&g_posture_info)[k_ai_posture_count];
-extern short(&g_variant_counts)[k_activity_count];
+extern int16(&g_variant_counts)[k_activity_count];
 
 extern void __cdecl activities_initialize();
