+++ game/source/ai/actor_dynamic_firing_position.hpp
@@ -8,10 +8,10 @@
 struct firing_position_definition
 {
 	c_ai_point3d position;
-	short flags;
+	int16 flags;
 	uint16 posture_flags;
-	short area_index;
-	short cluster_index;
+	int16 area_index;
+	int16 cluster_index;
 	c_sector_ref sector_ref;
 	real_euler_angles2d normal;
 	real32 yaw;
@@ -21,26 +21,26 @@ static_assert(sizeof(firing_position_definition) == 0x28);
 struct dynamic_firing_set_datum :
 	s_datum_header
 {
-	short type;
-	long support_object_index;
-	short position_count;
-	short num_members;
+	int16 type;
+	int32 support_object_index;
+	int16 position_count;
+	int16 num_members;
 	c_ai_point3d position;
 	real_vector3d up;
 	real_vector3d forward;
 	c_sector_ref sector_ref;
 	bool state_valid;
-	long last_update_time;
+	int32 last_update_time;
 	c_static_flags<32> initialized;
 	firing_position_definition firing_positions[32];
-	short firing_position_types[32];
+	int16 firing_position_types[32];
 };
 static_assert(sizeof(dynamic_firing_set_datum) == 0x584);
 
 struct dynamic_firing_set_iterator
 {
 	c_data_iterator<dynamic_firing_set_datum> iterator;
-	long index;
+	int32 index;
 };
 static_assert(sizeof(dynamic_firing_set_iterator) == 0x14);
